+++ changelog.md
@@ -17,6 +17,7 @@ New
 * Dump: new class developed and tested
 * Password: digest method added
 * User_Agent: detect_operating_system method added
+* User_Agent: list_operating_systems method added
 
 Since php-library v1.5.0 (6-Apr-2018)
 =======